+++ src/main/java/org/intellij/sdk/project/model/MyToolWindow.java
@@ -13,6 +13,11 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import javax.swing.*;
 import javax.swing.tree.DefaultTreeCellRenderer;
 import javax.swing.tree.DefaultTreeModel;
+import org.intellij.sdk.project.model.services.ComputeChildrenService;
+import org.intellij.sdk.project.model.services.DiffService;
+import org.intellij.sdk.project.model.services.NodeNavigatorService;
+import org.intellij.sdk.project.model.services.ParserService;
+import org.intellij.sdk.project.model.services.PersistencyService;
 import org.intellij.sdk.project.model.xnodes.XTestCompositeNode;
 import org.jetbrains.annotations.NotNull;
 import com.intellij.openapi.components.ServiceManager;
