+++ src/main/java/org/intellij/sdk/project/model/services/ParserService.java
@@ -1,4 +1,4 @@
-package org.intellij.sdk.project.model;
+package org.intellij.sdk.project.model.services;
 
 import java.util.ArrayList;
 import java.util.Arrays;
