+++ src/main/java/org/intellij/sdk/project/model/services/SnapService.java
@@ -1,7 +1,8 @@
 // Copyright 2000-2020 JetBrains s.r.o. and other contributors. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 
-package org.intellij.sdk.project.model;
+package org.intellij.sdk.project.model.services;
 
+import org.intellij.sdk.project.model.MyToolWindow;
 import org.jetbrains.annotations.NotNull;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.wm.ToolWindow;
