+++ src/main/resources/META-INF/plugin.xml
@@ -33,7 +33,7 @@
     <vendor url="https://plugins.jetbrains.com">IntelliJ Platform SDK</vendor>
     <extensions defaultExtensionNs="com.intellij">
         <toolWindow id="Armadillo" secondary="true" icon="AllIcons.General.Modified" anchor="bottom"
-                    factoryClass="org.intellij.sdk.project.model.SnapService"/>
-        <applicationService serviceImplementation="org.intellij.sdk.project.model.PersistencyService" serviceInterface="org.intellij.sdk.project.model.PersistencyService"/>
+                    factoryClass="org.intellij.sdk.project.model.services.SnapService"/>
+        <applicationService serviceImplementation="org.intellij.sdk.project.model.services.PersistencyService" serviceInterface="org.intellij.sdk.project.model.services.PersistencyService"/>
     </extensions>
 </idea-plugin>
